Network.AWS.MarketplaceAnalytics

Description

Provides AWS Marketplace business intelligence data on-demand.

Service Configuration

marketplaceAnalytics :: Service

API version 2015-07-01 of the Amazon Marketplace Commerce Analytics SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MarketplaceAnalytics.

MarketplaceCommerceAnalyticsException

_MarketplaceCommerceAnalyticsException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when an internal service error occurs.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

StartSupportDataExport

module Network.AWS.MarketplaceAnalytics.StartSupportDataExport

GenerateDataSet

module Network.AWS.MarketplaceAnalytics.GenerateDataSet

Types

DataSetType

data DataSetType

Constructors

CustomerProfileByGeography

CustomerProfileByIndustry

CustomerProfileByRevenue

CustomerSubscriberAnnualSubscriptions

CustomerSubscriberHourlyMonthlySubscriptions

DailyBusinessCanceledProductSubscribers

DailyBusinessFees

DailyBusinessFreeTrialConversions

DailyBusinessNewInstances

DailyBusinessNewProductSubscribers

DailyBusinessUsageByInstanceType

DisbursedAmountByAgeOfDisbursedFunds

DisbursedAmountByAgeOfUncollectedFunds

DisbursedAmountByCustomerGeo

DisbursedAmountByInstanceHours

DisbursedAmountByProduct

DisbursedAmountByProductWithUncollectedFunds

MonthlyRevenueAnnualSubscriptions

MonthlyRevenueBillingAndRevenueData

SalesCompensationBilledRevenue

UsSalesAndUseTaxRecords

SupportDataSetType

data SupportDataSetType

Constructors

CustomerSupportContactsData

TestCustomerSupportContactsData