Network.AWS.MarketplaceAnalytics.GenerateDataSet

Description

Given a data set type and data set publication date, asynchronously publishes the requested data set to the specified S3 bucket and notifies the specified SNS topic once the data is available. Returns a unique request identifier that can be used to correlate requests with notifications from the SNS topic. Data sets will be published in comma-separated values (CSV) format with the file name {data_set_type}_YYYY-MM-DD.csv. If a file with the same name already exists (e.g. if the same data set is requested twice), the original file will be overwritten by the new file. Requires a Role with an attached permissions policy providing Allow permissions for the following actions: s3:PutObject, s3:GetBucketLocation, sns:GetTopicAttributes, sns:Publish, iam:GetRolePolicy.

Creating a Request

generateDataSet

Arguments

:: DataSetType
gdsDataSetType

-> UTCTime
gdsDataSetPublicationDate

-> Text
gdsRoleNameARN

-> Text
gdsDestinationS3BucketName

-> Text
gdsSnsTopicARN

-> GenerateDataSet

Creates a value of GenerateDataSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsCustomerDefinedValues - (Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file. These key-value pairs can be used to correlated responses with tracking information from other systems.

gdsDestinationS3Prefix - (Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefixmydatasets", the output file "outputfile" would be published to "s3:mybucketmyprefixmydatasetsoutputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

gdsDataSetType - The desired data set type. * customer_subscriber_hourly_monthly_subscriptions From 2014-07-21 to present: Available daily by 5:00 PM Pacific Time. * customer_subscriber_annual_subscriptions From 2014-07-21 to present: Available daily by 5:00 PM Pacific Time. * daily_business_usage_by_instance_type From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_fees From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_free_trial_conversions From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_new_instances From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_new_product_subscribers From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_canceled_product_subscribers From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * monthly_revenue_billing_and_revenue_data From 2015-02 to 2017-06: Available monthly on the 4th day of the month by 5:00pm Pacific Time. Data includes metered transactions (e.g. hourly) from two months prior. From 2017-07 to present: Available monthly on the 15th day of the month by 5:00pm Pacific Time. Data includes metered transactions (e.g. hourly) from one month prior. * monthly_revenue_annual_subscriptions From 2015-02 to 2017-06: Available monthly on the 4th day of the month by 5:00pm Pacific Time. Data includes up-front software charges (e.g. annual) from one month prior. From 2017-07 to present: Available monthly on the 15th day of the month by 5:00pm Pacific Time. Data includes up-front software charges (e.g. annual) from one month prior. * disbursed_amount_by_product From 2015-01-26 to present: Available every 30 days by 5:00 PM Pacific Time. * disbursed_amount_by_product_with_uncollected_funds From 2012-04-19 to 2015-01-25: Available every 30 days by 5:00 PM Pacific Time. From 2015-01-26 to present: This data set was split into three data sets: disbursed_amount_by_product, disbursed_amount_by_age_of_uncollected_funds, and disbursed_amount_by_age_of_disbursed_funds. * disbursed_amount_by_instance_hours From 2012-09-04 to present: Available every 30 days by 5:00 PM Pacific Time. * disbursed_amount_by_customer_geo From 2012-04-19 to present: Available every 30 days by 5:00 PM Pacific Time. * disbursed_amount_by_age_of_uncollected_funds From 2015-01-26 to present: Available every 30 days by 5:00 PM Pacific Time. * disbursed_amount_by_age_of_disbursed_funds From 2015-01-26 to present: Available every 30 days by 5:00 PM Pacific Time. * customer_profile_by_industry From 2015-10-01 to 2017-06-29: Available daily by 5:00 PM Pacific Time. From 2017-06-30 to present: This data set is no longer available. * customer_profile_by_revenue From 2015-10-01 to 2017-06-29: Available daily by 5:00 PM Pacific Time. From 2017-06-30 to present: This data set is no longer available. * customer_profile_by_geography From 2015-10-01 to 2017-06-29: Available daily by 5:00 PM Pacific Time. From 2017-06-30 to present: This data set is no longer available. * sales_compensation_billed_revenue From 2016-12 to 2017-06: Available monthly on the 4th day of the month by 5:00pm Pacific Time. Data includes metered transactions (e.g. hourly) from two months prior, and up-front software charges (e.g. annual) from one month prior. From 2017-06 to present: Available monthly on the 15th day of the month by 5:00pm Pacific Time. Data includes metered transactions (e.g. hourly) from one month prior, and up-front software charges (e.g. annual) from one month prior. * us_sales_and_use_tax_records From 2017-02-15 to present: Available monthly on the 15th day of the month by 5:00 PM Pacific Time.

gdsDataSetPublicationDate - The date a data set was published. For daily data sets, provide a date with day-level granularity for the desired day. For weekly data sets, provide a date with day-level granularity within the desired week (the day value will be ignored). For monthly data sets, provide a date with month-level granularity for the desired month (the day value will be ignored).

gdsRoleNameARN - The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

gdsDestinationS3BucketName - The name (friendly name, not ARN) of the destination S3 bucket.

gdsSnsTopicARN - Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

data GenerateDataSet

Container for the parameters to the GenerateDataSet operation.

See: generateDataSet smart constructor.

Request Lenses

gdsCustomerDefinedValues :: Lens' GenerateDataSet (HashMap Text Text)

(Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file. These key-value pairs can be used to correlated responses with tracking information from other systems.

gdsDestinationS3Prefix :: Lens' GenerateDataSet (Maybe Text)

(Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefixmydatasets", the output file "outputfile" would be published to "s3:mybucketmyprefixmydatasetsoutputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

gdsDataSetType :: Lens' GenerateDataSet DataSetType

The desired data set type. * customer_subscriber_hourly_monthly_subscriptions From 2014-07-21 to present: Available daily by 5:00 PM Pacific Time. * customer_subscriber_annual_subscriptions From 2014-07-21 to present: Available daily by 5:00 PM Pacific Time. * daily_business_usage_by_instance_type From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_fees From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_free_trial_conversions From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_new_instances From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_new_product_subscribers From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * daily_business_canceled_product_subscribers From 2015-01-26 to present: Available daily by 5:00 PM Pacific Time. * monthly_revenue_billing_and_revenue_data From 2015-02 to 2017-06: Available monthly on the 4th day of the month by 5:00pm Pacific Time. Data includes metered transactions (e.g. hourly) from two months prior. From 2017-07 to present: Available monthly on the 15th day of the month by 5:00pm Pacific Time. Data includes metered transactions (e.g. hourly) from one month prior. * monthly_revenue_annual_subscriptions From 2015-02 to 2017-06: Available monthly on the 4th day of the month by 5:00pm Pacific Time. Data includes up-front software charges (e.g. annual) from one month prior. From 2017-07 to present: Available monthly on the 15th day of the month by 5:00pm Pacific Time. Data includes up-front software charges (e.g. annual) from one month prior. * disbursed_amount_by_product From 2015-01-26 to present: Available every 30 days by 5:00 PM Pacific Time. * disbursed_amount_by_product_with_uncollected_funds From 2012-04-19 to 2015-01-25: Available every 30 days by 5:00 PM Pacific Time. From 2015-01-26 to present: This data set was split into three data sets: disbursed_amount_by_product, disbursed_amount_by_age_of_uncollected_funds, and disbursed_amount_by_age_of_disbursed_funds. * disbursed_amount_by_instance_hours From 2012-09-04 to present: Available every 30 days by 5:00 PM Pacific Time. * disbursed_amount_by_customer_geo From 2012-04-19 to present: Available every 30 days by 5:00 PM Pacific Time. * disbursed_amount_by_age_of_uncollected_funds From 2015-01-26 to present: Available every 30 days by 5:00 PM Pacific Time. * disbursed_amount_by_age_of_disbursed_funds From 2015-01-26 to present: Available every 30 days by 5:00 PM Pacific Time. * customer_profile_by_industry From 2015-10-01 to 2017-06-29: Available daily by 5:00 PM Pacific Time. From 2017-06-30 to present: This data set is no longer available. * customer_profile_by_revenue From 2015-10-01 to 2017-06-29: Available daily by 5:00 PM Pacific Time. From 2017-06-30 to present: This data set is no longer available. * customer_profile_by_geography From 2015-10-01 to 2017-06-29: Available daily by 5:00 PM Pacific Time. From 2017-06-30 to present: This data set is no longer available. * sales_compensation_billed_revenue From 2016-12 to 2017-06: Available monthly on the 4th day of the month by 5:00pm Pacific Time. Data includes metered transactions (e.g. hourly) from two months prior, and up-front software charges (e.g. annual) from one month prior. From 2017-06 to present: Available monthly on the 15th day of the month by 5:00pm Pacific Time. Data includes metered transactions (e.g. hourly) from one month prior, and up-front software charges (e.g. annual) from one month prior. * us_sales_and_use_tax_records From 2017-02-15 to present: Available monthly on the 15th day of the month by 5:00 PM Pacific Time.

gdsDataSetPublicationDate :: Lens' GenerateDataSet UTCTime

The date a data set was published. For daily data sets, provide a date with day-level granularity for the desired day. For weekly data sets, provide a date with day-level granularity within the desired week (the day value will be ignored). For monthly data sets, provide a date with month-level granularity for the desired month (the day value will be ignored).

gdsRoleNameARN :: Lens' GenerateDataSet Text

The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

gdsDestinationS3BucketName :: Lens' GenerateDataSet Text

The name (friendly name, not ARN) of the destination S3 bucket.

gdsSnsTopicARN :: Lens' GenerateDataSet Text

Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

Destructuring the Response

generateDataSetResponse

Arguments

:: Int
gdsrsResponseStatus

-> GenerateDataSetResponse

Creates a value of GenerateDataSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsrsDataSetRequestId - A unique identifier representing a specific request to the GenerateDataSet operation. This identifier can be used to correlate a request with notifications from the SNS topic.

gdsrsResponseStatus - -- | The response status code.

data GenerateDataSetResponse

Container for the result of the GenerateDataSet operation.

See: generateDataSetResponse smart constructor.

Response Lenses

gdsrsDataSetRequestId :: Lens' GenerateDataSetResponse (Maybe Text)

A unique identifier representing a specific request to the GenerateDataSet operation. This identifier can be used to correlate a request with notifications from the SNS topic.

gdsrsResponseStatus :: Lens' GenerateDataSetResponse Int

- | The response status code.