Network.AWS.MarketplaceAnalytics.StartSupportDataExport

Description

Given a data set type and a from date, asynchronously publishes the requested customer support data to the specified S3 bucket and notifies the specified SNS topic once the data is available. Returns a unique request identifier that can be used to correlate requests with notifications from the SNS topic. Data sets will be published in comma-separated values (CSV) format with the file name {data_set_type}_YYYY-MM-DD'T'HH-mm-ss'Z'.csv. If a file with the same name already exists (e.g. if the same data set is requested twice), the original file will be overwritten by the new file. Requires a Role with an attached permissions policy providing Allow permissions for the following actions: s3:PutObject, s3:GetBucketLocation, sns:GetTopicAttributes, sns:Publish, iam:GetRolePolicy.

Creating a Request

startSupportDataExport

Arguments

:: SupportDataSetType
ssdeDataSetType

-> UTCTime
ssdeFromDate

-> Text
ssdeRoleNameARN

-> Text
ssdeDestinationS3BucketName

-> Text
ssdeSnsTopicARN

-> StartSupportDataExport

Creates a value of StartSupportDataExport with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssdeCustomerDefinedValues - (Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file.

ssdeDestinationS3Prefix - (Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefixmydatasets", the output file "outputfile" would be published to "s3:mybucketmyprefixmydatasetsoutputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

ssdeDataSetType - Specifies the data set type to be written to the output csv file. The data set types customer_support_contacts_data and test_customer_support_contacts_data both result in a csv file containing the following fields: Product Id, Product Code, Customer Guid, Subscription Guid, Subscription Start Date, Organization, AWS Account Id, Given Name, Surname, Telephone Number, Email, Title, Country Code, ZIP Code, Operation Type, and Operation Time. * customer_support_contacts_data Customer support contact data. The data set will contain all changes (Creates, Updates, and Deletes) to customer support contact data from the date specified in the from_date parameter. * test_customer_support_contacts_data An example data set containing static test data in the same format as customer_support_contacts_data

ssdeFromDate - The start date from which to retrieve the data set in UTC. This parameter only affects the customer_support_contacts_data data set type.

ssdeRoleNameARN - The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

ssdeDestinationS3BucketName - The name (friendly name, not ARN) of the destination S3 bucket.

ssdeSnsTopicARN - Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

data StartSupportDataExport

Container for the parameters to the StartSupportDataExport operation.

See: startSupportDataExport smart constructor.

Request Lenses

ssdeCustomerDefinedValues :: Lens' StartSupportDataExport (HashMap Text Text)

(Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file.

ssdeDestinationS3Prefix :: Lens' StartSupportDataExport (Maybe Text)

(Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefixmydatasets", the output file "outputfile" would be published to "s3:mybucketmyprefixmydatasetsoutputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

ssdeDataSetType :: Lens' StartSupportDataExport SupportDataSetType

Specifies the data set type to be written to the output csv file. The data set types customer_support_contacts_data and test_customer_support_contacts_data both result in a csv file containing the following fields: Product Id, Product Code, Customer Guid, Subscription Guid, Subscription Start Date, Organization, AWS Account Id, Given Name, Surname, Telephone Number, Email, Title, Country Code, ZIP Code, Operation Type, and Operation Time. * customer_support_contacts_data Customer support contact data. The data set will contain all changes (Creates, Updates, and Deletes) to customer support contact data from the date specified in the from_date parameter. * test_customer_support_contacts_data An example data set containing static test data in the same format as customer_support_contacts_data

ssdeFromDate :: Lens' StartSupportDataExport UTCTime

The start date from which to retrieve the data set in UTC. This parameter only affects the customer_support_contacts_data data set type.

ssdeRoleNameARN :: Lens' StartSupportDataExport Text

The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

ssdeDestinationS3BucketName :: Lens' StartSupportDataExport Text

The name (friendly name, not ARN) of the destination S3 bucket.

ssdeSnsTopicARN :: Lens' StartSupportDataExport Text

Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

Destructuring the Response

startSupportDataExportResponse

Arguments

:: Int
ssdersResponseStatus

-> StartSupportDataExportResponse

Creates a value of StartSupportDataExportResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssdersDataSetRequestId - A unique identifier representing a specific request to the StartSupportDataExport operation. This identifier can be used to correlate a request with notifications from the SNS topic.

ssdersResponseStatus - -- | The response status code.

data StartSupportDataExportResponse

Container for the result of the StartSupportDataExport operation.

See: startSupportDataExportResponse smart constructor.

Response Lenses

ssdersDataSetRequestId :: Lens' StartSupportDataExportResponse (Maybe Text)

A unique identifier representing a specific request to the StartSupportDataExport operation. This identifier can be used to correlate a request with notifications from the SNS topic.

ssdersResponseStatus :: Lens' StartSupportDataExportResponse Int

- | The response status code.